requirements.txt
click==8.1.3
colorama==0.4.6
document==1.0
et-xmlfile==1.1.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
openpyxl==3.1.2
psycopg2==2.9.5
python-docx==0.8.11
SQLAlchemy==2.0.6
typing_extensions==4.5.0
Werkzeug==2.2.3
XlsxWriter==3.0.9
